[Forms Stuff Refactoring] from lib to app/forms
