+++ app/forms/qae_2014_forms.rb
@@ -1,4 +1,3 @@
-require 'qae_form_builder'
 require 'qae_2014_forms/innovation'
 require 'qae_2014_forms/international_trade'
 require 'qae_2014_forms/sustainable_development'
